backend/requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.3
python-dotenv==1.0.1
rdflib==7.1.1
requests==2.32.3
SPARQLWrapper==2.0.0
urllib3==2.2.3
Flask-Cors==5.0.0